refactor: 카페 단건조회 서비스 메소드 Query서비스단으로 변경
