+++ src/main/java/com/example/demo/controller/CafeController.java
@@ -33,7 +33,7 @@ public class CafeController {
 
 	@GetMapping("/cafe/{cafeId}")
 	public ResponseEntity<CafeResponse> cafe(@PathVariable Long cafeId) {
-		CafeResponse response = cafeService.findCafeById(cafeId);
+		CafeResponse response = cafeQueryService.findCafeById(cafeId);
 		return new ResponseEntity<>(response, HttpStatus.OK);
 	}
