+++ src/main/java/com/example/demo/service/CafeQueryService.java
@@ -1,11 +1,13 @@
 package com.example.demo.service;
 
+import com.example.demo.dto.CafeResponse;
 import com.example.demo.dto.CafeSearchRequest;
 import com.example.demo.dto.CafeSearchResponse;
 import com.example.demo.dto.PagedResponse;
 
 public interface CafeQueryService {
 
-	public PagedResponse<CafeSearchResponse> searchWithPagingByDynamicFilter(CafeSearchRequest request);
+	PagedResponse<CafeSearchResponse> searchWithPagingByDynamicFilter(CafeSearchRequest request);
 
+	CafeResponse findCafeById(Long cafeId);
 }
