+++ src/main/java/com/example/demo/service/CafeService.java
@@ -1,8 +1,4 @@
 package com.example.demo.service;
 
-import com.example.demo.dto.CafeResponse;
-
 public interface CafeService {
-
-	CafeResponse findCafeById(Long cafeId);
 }
